deps/v8/test/inspector/debugger/async-function-step-out-optimized-expected.txt

stepOut async function

Running test: testTrivial
Check that we have proper async stack at return
bar (testTrivial.js:30:8)
-- async function --
bar (testTrivial.js:29:22)
foo (testTrivial.js:25:14)
-- async function --
foo (testTrivial.js:24:22)
test (testTrivial.js:20:14)
-- async function --
test (testTrivial.js:19:22)
(anonymous) (:0:0)

foo (testTrivial.js:26:6)
-- async function --
foo (testTrivial.js:24:22)
test (testTrivial.js:20:14)
-- async function --
test (testTrivial.js:19:22)
(anonymous) (:0:0)

test (testTrivial.js:21:6)
-- async function --
test (testTrivial.js:19:22)
(anonymous) (:0:0)


Running test: testStepOutPrecision
Check that stepOut go to resumed outer generator
bar (testStepOutPrecision.js:63:8)
-- async function --
bar (testStepOutPrecision.js:62:22)
foo (testStepOutPrecision.js:57:14)
-- async function --
foo (testStepOutPrecision.js:56:22)
test (testStepOutPrecision.js:50:14)
-- async function --
test (testStepOutPrecision.js:49:14)
(anonymous) (:0:0)

foo (testStepOutPrecision.js:58:8)
-- async function --
foo (testStepOutPrecision.js:56:22)
test (testStepOutPrecision.js:50:14)
-- async function --
test (testStepOutPrecision.js:49:14)
(anonymous) (:0:0)

test (testStepOutPrecision.js:51:8)
-- async function --
test (testStepOutPrecision.js:49:14)
(anonymous) (:0:0)

floodWithTimeouts (testStepOutPrecision.js:42:15)
-- setTimeout --
floodWithTimeouts (testStepOutPrecision.js:43:10)
-- setTimeout --
floodWithTimeouts (testStepOutPrecision.js:43:10)
-- setTimeout --
floodWithTimeouts (testStepOutPrecision.js:43:10)
-- setTimeout --
floodWithTimeouts (testStepOutPrecision.js:43:10)
test (testStepOutPrecision.js:48:8)
(anonymous) (:0:0)

test (testStepOutPrecision.js:52:8)
-- async function --
test (testStepOutPrecision.js:49:14)
(anonymous) (:0:0)

floodWithTimeouts (testStepOutPrecision.js:42:15)
-- setTimeout --
floodWithTimeouts (testStepOutPrecision.js:43:10)
-- setTimeout --
floodWithTimeouts (testStepOutPrecision.js:43:10)
-- setTimeout --
floodWithTimeouts (testStepOutPrecision.js:43:10)
-- setTimeout --
floodWithTimeouts (testStepOutPrecision.js:43:10)
-- setTimeout --
floodWithTimeouts (testStepOutPrecision.js:43:10)
test (testStepOutPrecision.js:48:8)
(anonymous) (:0:0)


Running test: testStepIntoAtReturn
Check that stepInto at return go to resumed outer generator
bar (testStepIntoAtReturn.js:95:8)
-- async function --
bar (testStepIntoAtReturn.js:94:22)
foo (testStepIntoAtReturn.js:90:14)
-- async function --
foo (testStepIntoAtReturn.js:89:22)
test (testStepIntoAtReturn.js:84:14)
-- async function --
test (testStepIntoAtReturn.js:83:14)
(anonymous) (:0:0)

bar (testStepIntoAtReturn.js:96:6)
-- async function --
bar (testStepIntoAtReturn.js:94:22)
foo (testStepIntoAtReturn.js:90:14)
-- async function --
foo (testStepIntoAtReturn.js:89:22)
test (testStepIntoAtReturn.js:84:14)
-- async function --
test (testStepIntoAtReturn.js:83:14)
(anonymous) (:0:0)

foo (testStepIntoAtReturn.js:91:6)
-- async function --
foo (testStepIntoAtReturn.js:89:22)
test (testStepIntoAtReturn.js:84:14)
-- async function --
test (testStepIntoAtReturn.js:83:14)
(anonymous) (:0:0)

test (testStepIntoAtReturn.js:85:8)
-- async function --
test (testStepIntoAtReturn.js:83:14)
(anonymous) (:0:0)

test (testStepIntoAtReturn.js:86:6)
-- async function --
test (testStepIntoAtReturn.js:83:14)
(anonymous) (:0:0)

floodWithTimeouts (testStepIntoAtReturn.js:76:15)
-- setTimeout --
floodWithTimeouts (testStepIntoAtReturn.js:77:10)
-- setTimeout --
floodWithTimeouts (testStepIntoAtReturn.js:77:10)
test (testStepIntoAtReturn.js:82:8)
(anonymous) (:0:0)


Running test: testStepOverAtReturn
Check that stepOver at return go to resumed outer generator
bar (testStepIntoAtReturn.js:126:8)
-- async function --
bar (testStepIntoAtReturn.js:125:22)
foo (testStepIntoAtReturn.js:121:14)
-- async function --
foo (testStepIntoAtReturn.js:120:22)
test (testStepIntoAtReturn.js:115:14)
-- async function --
test (testStepIntoAtReturn.js:114:14)
(anonymous) (:0:0)

bar (testStepIntoAtReturn.js:127:6)
-- async function --
bar (testStepIntoAtReturn.js:125:22)
foo (testStepIntoAtReturn.js:121:14)
-- async function --
foo (testStepIntoAtReturn.js:120:22)
test (testStepIntoAtReturn.js:115:14)
-- async function --
test (testStepIntoAtReturn.js:114:14)
(anonymous) (:0:0)

foo (testStepIntoAtReturn.js:122:6)
-- async function --
foo (testStepIntoAtReturn.js:120:22)
test (testStepIntoAtReturn.js:115:14)
-- async function --
test (testStepIntoAtReturn.js:114:14)
(anonymous) (:0:0)

test (testStepIntoAtReturn.js:116:8)
-- async function --
test (testStepIntoAtReturn.js:114:14)
(anonymous) (:0:0)

test (testStepIntoAtReturn.js:117:6)
-- async function --
test (testStepIntoAtReturn.js:114:14)
(anonymous) (:0:0)

floodWithTimeouts (testStepIntoAtReturn.js:107:15)
-- setTimeout --
floodWithTimeouts (testStepIntoAtReturn.js:108:10)
-- setTimeout --
floodWithTimeouts (testStepIntoAtReturn.js:108:10)
test (testStepIntoAtReturn.js:113:8)
(anonymous) (:0:0)


Running test: testStepOutFromNotAwaitedCall
Checks stepOut from not awaited call
bar (testStepIntoAtReturn.js:160:8)
-- async function --
bar (testStepIntoAtReturn.js:159:22)
foo (testStepIntoAtReturn.js:154:8)
-- async function --
foo (testStepIntoAtReturn.js:153:22)
test (testStepIntoAtReturn.js:146:14)
-- async function --
test (testStepIntoAtReturn.js:145:14)
(anonymous) (:0:0)

test (testStepIntoAtReturn.js:147:8)
-- async function --
test (testStepIntoAtReturn.js:145:14)
(anonymous) (:0:0)

floodWithTimeouts (testStepIntoAtReturn.js:138:15)
-- setTimeout --
floodWithTimeouts (testStepIntoAtReturn.js:139:10)
-- setTimeout --
floodWithTimeouts (testStepIntoAtReturn.js:139:10)
test (testStepIntoAtReturn.js:144:8)
(anonymous) (:0:0)
